fix: fix ts error
